backend/requirements.txt
fastapi
uvicorn[standard]
PyPDF2
google.generativeai
python-dotenv
PyMuPDF
python-multipart
